익명 관련 처리 제외 오류 수정

+++ board/views.py
@@ -36,20 +36,42 @@ def popular_posts(request):
 # -------------------- 내가 쓴 글 --------------------
 @login_required
 def my_posts(request):
-    posts = Post.objects.filter(user=request.user).order_by('-id')
-    return render(request, 'board/post_list.html', {'posts': posts, 'category': '내가 쓴 글'})
+    # 내가 작성한 글 (익명이든 아니든)
+    posts = Post.objects.filter(user=request.user)
+
+    # 내가 익명으로 작성한 글 (익명으로 작성한 글만 따로 필터링)
+    anonymous_posts = Post.objects.filter(user=request.user, is_anonymous=True)
+
+    # 두 리스트 합치기
+    all_posts = list(posts) + list(anonymous_posts)
+
+    return render(request, 'board/post_list.html', {'posts': all_posts})
+
 
 # -------------------- 내가 댓글 단 글 --------------------
 @login_required
 def my_comments(request):
-    posts = Post.objects.filter(comments__user=request.user).distinct().order_by('-id')
-    return render(request, 'board/post_list.html', {'posts': posts, 'category': '내가 댓글 단 글'})
+    # 내가 작성한 댓글과 익명 댓글을 포함한 댓글 목록 가져오기
+    comments = Comment.objects.filter(user=request.user)  # 내가 작성한 일반 댓글
+    anonymous_comments = Comment.objects.filter(user=None, post__comments__user=request.user)  # 내가 작성한 익명 댓글
+    
+    # 두 쿼리셋을 합침 (|는 OR 연산자)
+    combined_comments = comments | anonymous_comments
+    combined_comments = combined_comments.order_by('-created_at')  # 최신 댓글부터 정렬
+
+    return render(request, 'board/comment_list.html', {'comments': combined_comments})
+
+
+
 
 # -------------------- 내가 스크랩한 글 --------------------
 @login_required
 def my_scraps(request):
+    # 사용자가 스크랩한 게시글을 가져오기
     posts = Post.objects.filter(scraps=request.user).order_by('-id')
-    return render(request, 'board/post_list.html', {'posts': posts, 'category': '내 스크랩'})
+    
+    return render(request, 'board/scrap_list.html', {'scraps': posts, 'category': '내 스크랩'})
+
 
 def post_detail(request, pk):
     post = get_object_or_404(Post, pk=pk)
@@ -104,13 +126,11 @@ def post_create(request, category=None):
         }
     )
 
-# -------------------- 게시글 수정 --------------------
-@login_required
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
 
-    # 로그인한 사용자만 수정할 수 있도록 처리 (익명 글이라도 수정 권한 부여)
-    if request.user != post.user and post.user is not None:
+    # 수정하려는 글이 익명 글이라도 작성자는 자기 글을 수정할 수 있어야 함
+    if request.user != post.user and (post.user is not None or not post.is_anonymous):
         return redirect('board:post_detail', pk=pk)
 
     if request.method == 'POST':
@@ -133,15 +153,17 @@ def post_edit(request, pk):
         # 파일 처리 (새로운 파일을 올리면 기존 파일 교체)
         if request.FILES.get('file'):
             post.file = request.FILES.get('file')
-        if request.FILES.get('image'):
-            post.image = request.FILES.get('image')
+
+        # 파일 삭제 처리
+        if 'delete_file' in request.POST:
+            post.file.delete()
+            post.file = None
 
         # 게시글 수정 후 저장
         post.save()
 
         return redirect('board:post_detail', pk=pk)
 
-    # GET 요청일 때, 수정 페이지 렌더링
     return render(request, 'board/post_edit_page.html', {'post': post})
 
 # -------------------- 게시글 삭제 --------------------
@@ -149,8 +171,8 @@ def post_edit(request, pk):
 def post_delete(request, pk):
     post = get_object_or_404(Post, pk=pk)
 
-    # 게시글을 삭제하는 사용자 확인 (익명 글도 삭제 가능)
-    if request.user == post.user or post.user is None:
+    # 게시글 삭제 권한 확인: 익명 글이라도 작성자 본인만 삭제 가능
+    if request.user == post.user or (post.is_anonymous and post.user is None):
         category = post.category  # 게시글의 카테고리
         post.delete()
         
@@ -164,11 +186,16 @@ def post_delete(request, pk):
 @login_required
 def post_scrap_toggle(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    
+    # 스크랩 상태를 토글
     if request.user in post.scraps.all():
-        post.scraps.remove(request.user)
+        post.scraps.remove(request.user)  # 스크랩 취소
     else:
-        post.scraps.add(request.user)
-    return redirect('board:post_detail', pk=pk)
+        post.scraps.add(request.user)  # 스크랩
+
+    # 스크랩 상태를 처리한 후, 내가 스크랩한 글 목록 페이지로 리다이렉트
+    return redirect('board:my_scraps')
+
 
 # -------------------- 댓글 작성 --------------------
 @login_required
@@ -191,19 +218,28 @@ def comment_create(request, post_pk, parent_pk=None):
             parent=parent,
             is_anonymous=is_anonymous
         )
-        return redirect('board:post_detail', pk=post.pk)
-
+        
+        # 댓글 작성 후 내가 쓴 댓글 페이지로 리다이렉트
+        return redirect('board:my_comments')  # 내가 쓴 댓글 페이지로 리다이렉트
+    
+    # GET 요청인 경우에는 게시글 상세 페이지로 리다이렉트
     return redirect('board:post_detail', pk=post.pk)
 
 # -------------------- 댓글 좋아요 토글 --------------------
 @login_required
 def comment_like_toggle(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
+
     if request.user in comment.likes.all():
+        # 좋아요가 눌린 상태 -> 좋아요 취소
         comment.likes.remove(request.user)
     else:
+        # 좋아요가 눌리지 않은 상태 -> 좋아요 추가
         comment.likes.add(request.user)
+
+    # 댓글이 속한 게시글로 리다이렉트
     return redirect('board:post_detail', pk=comment.post.pk)
+
 
 # -------------------- 게시글 좋아요 토글 --------------------
 @login_required
